Validate the default value.
Just to be safe assume that the default value provided in the schema
may contain errors and run it through the conversion and validation
steps as any other value.

In "Schema.__init__", if a default value is provided (a dict) convert
it to a "MultiValueDict" to make "Schema.to_python" happy. Add tests
for the conversion.

Fix the tests for which i have been too lazy to care about the type
of the default value.

Adapt tests to the new semantic.