Remove "response_schema" from endpoints.
Schema's model isn't powerful enough to represent arbitrary data
structures. Any complete solution will be disproportionately complex
in regards to the problem that I'm trying to solve here.

I think that not providing the formalized response specification is
better than providing an incomplete or misleading one. An informal
specification can be provided in the endpoint's description.